commands/config: Rename list command to describe
