+++ internal/commands/config/root.go
@@ -4,9 +4,9 @@ import (
 	"github.com/urfave/cli/v3"
 
 	"github.com/foohq/foojank/internal/actions"
+	"github.com/foohq/foojank/internal/commands/config/describe"
 	"github.com/foohq/foojank/internal/commands/config/edit"
 	confinit "github.com/foohq/foojank/internal/commands/config/init"
-	"github.com/foohq/foojank/internal/commands/config/list"
 )
 
 func NewCommand() *cli.Command {
@@ -15,8 +15,8 @@ func NewCommand() *cli.Command {
 		Usage: "Manage configuration",
 		Commands: []*cli.Command{
 			confinit.NewCommand(),
+			describe.NewCommand(),
 			edit.NewCommand(),
-			list.NewCommand(),
 		},
 		CommandNotFound: actions.CommandNotFound,
 		OnUsageError:    actions.UsageError,
